Priority Donation: Done for conditions too
This greatly simplifies the implementation for conditions. The
priority handling is completely handled by the semaphore used for each
condition.

I was skeptical about the base implementation being too complicated,
and this confirms my suspicion. It's not obvious to me right now why a
single semaphore was not used originally. Maybe there was no semaphore
waiting list in baseline pintos?